gcp/requirements.txt
tensorflow==2.16.1
google-cloud-storage==1.16.1
Pillow==6.0.0